moved anything that modifies spec into utils
